Create CannonBall class, add CANNON_BALL type in EntityType and add a instance in EntityFactoryImpl, quick parameter change in function of Vector2d

+++ src/main/java/it/unibo/cicciopier/utility/Vector2d.java
@@ -99,7 +99,7 @@ public class Vector2d {
      *
      * @param length desired to be
      */
-    public void setMagnitude(final int length) {
+    public void setMagnitude(final double length) {
         this.normalize();
         this.scale(length);
     }
